remove warnings

+++ core/src/Scene.cpp
@@ -11,7 +11,7 @@ using namespace TracerX::core;
 
 int Scene::addTexture(const Image& texture, const std::string& name)
 {
-    int index = this->textures.size();
+    int index = (int)this->textures.size();
     this->textures.push_back(texture);
     this->textureNames.push_back(name.empty() ? "Untitled texture " + std::to_string(index) : name);
     return index;
@@ -19,7 +19,7 @@ int Scene::addTexture(const Image& texture, const std::string& name)
 
 int Scene::addMaterial(const Material& material, const std::string& name)
 {
-    int index = this->materials.size();
+    int index = (int)this->materials.size();
     this->materials.push_back(material);
     this->materialNames.push_back(name.empty() ? "Untitled material " + std::to_string(index) : name);
     return index;
@@ -27,7 +27,7 @@ int Scene::addMaterial(const Material& material, const std::string& name)
 
 int Scene::addMesh(const Mesh& mesh, const std::string& name)
 {
-    int index = this->meshes.size();
+    int index = (int)this->meshes.size();
     this->meshes.push_back(mesh);
     this->meshNames.push_back(name.empty() ? "Untitled mesh " + std::to_string(index) : name);
     this->buildBVH(this->meshes.back());
@@ -62,7 +62,7 @@ void Scene::buildBVH(Mesh& mesh)
         triangleConverter);
 
     // Convert to our format
-    mesh.nodeOffset = this->bvh.size();
+    mesh.nodeOffset = (int)this->bvh.size();
     this->bvh.reserve(mesh.nodeOffset + bvh.getNodes().size());
     for (const FastBVH::Node<float>& node : bvh.getNodes())
     {
